+++ src/builder/__tests__/extractOverrides.ts
@@ -1,7 +1,7 @@
 import {BuildTimeConfig, Overrides} from '../types';
 
-export function extractOverrides<Preset, Result, Parameters extends any[]>(
-    config?: BuildTimeConfig<Preset, unknown, Result, Parameters>,
-): Overrides<Preset> {
+export function extractOverrides<Origin, Result = Origin, PostBuildResult = Result>(
+    config?: BuildTimeConfig<Origin, Result, PostBuildResult>,
+): Overrides<Origin> {
     return config?.overrides ?? {};
 }