+++ src/builder/builder.ts
@@ -21,29 +21,33 @@ import {extractOverrides} from './__tests__/extractOverrides';
 import {extractFieldsConfiguration} from './utils/extractFieldsConfiguration';
 import {extractInitialParameters} from './utils/extractInitialParameters';
 
-function createBuilder<Origin, Fields = Origin, Trait extends string = string, InitialParameters extends any[] = never>(
-    config: BuilderConfiguration<Origin, Fields, Trait, InitialParameters>,
-): Builder<Origin, Fields, Trait, InitialParameters> {
+function createBuilder<Origin, Result = Origin, Trait extends string = string, InitialParameters extends any[] = never>(
+    config: BuilderConfiguration<Origin, Result, Trait, InitialParameters>,
+): Builder<Origin, Result, Trait, InitialParameters> {
     const fieldsConfiguration = extractFieldsConfiguration(config.fields);
 
     const traits: TraitsConfiguration<Origin, Trait> | undefined = config.traits;
-    const postBuild: ((x: Origin) => Fields) | undefined = config.postBuild;
+    const postBuild: ((x: Origin) => Result) | undefined = config.postBuild;
 
     let definedIterators: IteratorsConfiguration<Origin> | null = null;
     let previousBuildFields: Origin | undefined;
     let fieldsConfigurationGenerator: FieldsConfigurationGenerator<Origin> | null = null;
 
-    const mapFieldsWithOverrides = <Fields = Origin, MapperBuild = Fields>(
+    const mapFieldsWithOverrides = <Fields = Origin, PostBuildResult = Result>(
         fields: FieldsConfiguration<Fields>,
-        buildConfig?: BuildTimeConfig<Fields, Trait, MapperBuild, InitialParameters>,
+        buildConfig?: BuildTimeConfig<Fields, Result, PostBuildResult, Trait>,
     ) => {
+        if (!isPlainObject(fields)) {
+            return fields;
+        }
+
         const buildOverrides = extractOverrides(buildConfig);
         const buildTraitsOverrides = extractTraits(buildConfig).reduce<Overrides<Fields>>(
             (traitsOverrides, traitKey) => {
                 if (!traits?.[traitKey]) {
                     console.warn(`Trait "${String(traitKey)}" is not specified in buildConfig!`);
                 }
-                const traitsConfig = traits ? traits[traitKey] : ({} as TraitsConfiguration<Origin>);
+                const traitsConfig = traits ? traits[traitKey] : ({} as TraitsConfiguration<Fields>);
                 const currentTraitOverrides = traitsConfig.overrides ?? {};
                 return deepMerge(traitsOverrides, currentTraitOverrides);
             },
@@ -100,7 +104,7 @@ function createBuilder<Origin, Fields = Origin, Trait extends string = string, I
         }
 
         if (isIterator(field)) {
-            return field.next().value;
+            return field.next().value as Value;
         }
 
         if (isClassInstance<Value>(field)) {
@@ -118,8 +122,8 @@ function createBuilder<Origin, Fields = Origin, Trait extends string = string, I
         return field;
     };
 
-    const initFieldsGenerator = <MapperBuild = Fields>(
-        buildConfig?: BuildTimeConfig<Origin, Trait, MapperBuild, InitialParameters>,
+    const initFieldsGenerator = <PostBuildResult = Result>(
+        buildConfig?: BuildTimeConfig<Origin, Result, PostBuildResult, Trait, InitialParameters>,
     ) => {
         if (!fieldsConfiguration.originFieldsGenerator) {
             return;
@@ -129,10 +133,8 @@ function createBuilder<Origin, Fields = Origin, Trait extends string = string, I
         fieldsConfigurationGenerator = fieldsConfiguration.originFieldsGenerator(...initialParameters);
     };
 
-    const build = <MapperBuild = Fields>(
-        buildConfig?: BuildTimeConfig<Origin, Trait, MapperBuild, InitialParameters>,
-    ) => {
-        let fieldsForProcessing;
+    const build = <PostBuildResult = Result>(buildConfig?: BuildTimeConfig<Origin, Result, PostBuildResult, Trait>) => {
+        let fieldsForProcessing: FieldsConfiguration<Origin>;
 
         if (fieldsConfiguration.originFieldsFunction) {
             const iterationFields = fieldsConfiguration.originFieldsFunction(previousBuildFields);
@@ -142,7 +144,8 @@ function createBuilder<Origin, Fields = Origin, Trait extends string = string, I
                 throw Error("The fields GeneratorFunction isn't initialized!");
             }
 
-            const iterationFields = fieldsConfigurationGenerator.next(previousBuildFields).value;
+            const iterationFields = fieldsConfigurationGenerator.next(previousBuildFields)
+                .value as FieldsConfiguration<Origin>;
             fieldsForProcessing = deepMerge(iterationFields, extractIterators(iterationFields));
         } else {
             fieldsForProcessing = fieldsConfiguration.originFields;
@@ -153,23 +156,24 @@ function createBuilder<Origin, Fields = Origin, Trait extends string = string, I
         previousBuildFields = fields as Origin;
 
         const built = postBuild ? postBuild(fields as Origin) : fields;
-        return buildConfig?.postBuild ? buildConfig.postBuild(built as Origin) : (built as MapperBuild);
+        return buildConfig?.postBuild ? buildConfig.postBuild(built as Result) : (built as PostBuildResult);
     };
 
     return {
-        one: <Result = Fields>(buildConfig?: BuildTimeConfig<Origin, Trait, Result, InitialParameters>) => {
+        one: <PostBuildResult = Result>(buildConfig?: BuildTimeConfig<Origin, Result, PostBuildResult, Trait>) => {
             initFieldsGenerator(buildConfig);
             return build(buildConfig);
         },
-        many: <Result = Fields>(
+        many: <PostBuildResult = Result>(
             count: number,
-            buildConfig?: BuildTimeConfig<Origin, Trait, Result, InitialParameters>,
+            buildConfig?: BuildTimeConfig<Origin, Result, PostBuildResult, Trait>,
         ) => {
             initFieldsGenerator(buildConfig);
             return Array(count)
                 .fill(0)
                 .map(() => build(buildConfig));
         },
+        reset: () => {},
     };
 }