+++ src/builder/types.ts
@@ -3,68 +3,73 @@ import type {FieldsGenerator} from '../generators/generate';
 
 export type FunctionalGenerator<T> = () => T;
 
-export type FieldGenerator<T> = FunctionalGenerator<T> | Iterator<T, T>;
+export type FieldGenerator<T> = FunctionalGenerator<T> | Iterator<T, never | void>;
 
 export type FieldType<T> = T | FixedValue<T> | FieldGenerator<T> | FieldsConfiguration<T>;
 
-export type FieldsConfiguration<Result> = {
-    readonly [Key in keyof Result]: FieldType<Result[Key]>;
+export type FieldsConfiguration<Origin> = {
+    readonly [Key in keyof Origin]: FieldType<Origin[Key]>;
 };
 
-export type FieldsConfigurationFunction<FactoryResult> = (
-    prevBuild?: FactoryResult,
-) => FieldsConfiguration<FactoryResult>;
+export type FieldsConfigurationFunction<Origin> = (prevBuild?: Origin) => FieldsConfiguration<Origin>;
 
-export type FieldsConfigurationGenerator<FactoryResult> = Generator<
-    FieldsConfiguration<FactoryResult>,
-    FieldsConfiguration<FactoryResult>,
-    FactoryResult | undefined
+export type FieldsConfigurationGenerator<Origin> = Generator<
+    FieldsConfiguration<Origin>,
+    never | void,
+    Origin | undefined
 >;
 
-export type FieldsConfigurationGeneratorFunction<FactoryResult, InitialParameters extends any[] = never> = (
+export type FieldsConfigurationGeneratorFunction<Origin, InitialParameters extends any[] = never> = (
     ...initialParameters: InitialParameters
-) => FieldsConfigurationGenerator<FactoryResult>;
+) => FieldsConfigurationGenerator<Origin>;
 
-export type BuilderConfigurationFields<FactoryResult, InitialParameters extends any[] = never> =
-    | FieldsConfiguration<FactoryResult>
-    | FieldsConfigurationFunction<FactoryResult>
-    | FieldsGenerator<FactoryResult, InitialParameters>;
+export type BuilderConfigurationFields<Origin, InitialParameters extends any[] = never> =
+    | FieldsConfiguration<Origin>
+    | FieldsConfigurationFunction<Origin>
+    | FieldsGenerator<Origin, InitialParameters>;
 
 export type BuilderConfiguration<
-    FactoryResult,
-    PostBuildResult = FactoryResult,
+    Origin,
+    PostBuildResult = Origin,
     TraitName extends string = string,
     InitialParameters extends any[] = never,
 > = {
-    readonly fields: BuilderConfigurationFields<FactoryResult, InitialParameters>;
-    readonly traits?: TraitsConfiguration<FactoryResult, TraitName>;
-    readonly postBuild?: (x: FactoryResult) => PostBuildResult;
+    readonly fields: BuilderConfigurationFields<Origin, InitialParameters>;
+    readonly traits?: TraitsConfiguration<Origin, TraitName>;
+    readonly postBuild?: (x: Origin) => PostBuildResult;
 };
 
 export type Overrides<Result> = {
     [Key in keyof Result]?: FieldType<Result[Key]> | Overrides<Result[Key]>;
 };
 
-export type BuildTimeConfig<Result, Trait, MappedResult = Result, InitialParameters extends any[] = never> = {
-    overrides?: Overrides<Result>;
-    postBuild?: (builtThing: Result) => MappedResult;
-    traits?: Trait | Trait[];
+export type BuildTimeConfig<
+    Origin,
+    StaticPostBuildResult = Origin,
+    PostBuildResult = StaticPostBuildResult,
+    TraitName = string,
+    InitialParameters extends any[] = never,
+> = {
+    overrides?: Overrides<Origin>;
+    postBuild?: (builtThing: StaticPostBuildResult) => PostBuildResult;
+    traits?: TraitName | TraitName[];
     initialParameters?: InitialParameters;
 };
 
 export interface Builder<
-    FactoryResult,
-    PostBuildResult = FactoryResult,
+    Origin,
+    PostBuildResult = Origin,
     TraitName extends string = string,
     InitialParameters extends any[] = never,
 > {
     one<Result = PostBuildResult>(
-        buildTimeConfig?: BuildTimeConfig<FactoryResult, TraitName, Result, InitialParameters>,
+        buildTimeConfig?: BuildTimeConfig<Origin, PostBuildResult, Result, TraitName, InitialParameters>,
     ): Result;
     many<Result = PostBuildResult>(
         count: number,
-        buildTimeConfig?: BuildTimeConfig<FactoryResult, TraitName, Result, InitialParameters>,
+        buildTimeConfig?: BuildTimeConfig<Origin, PostBuildResult, Result, TraitName, InitialParameters>,
     ): Result[];
+    reset(): void;
 }
 
 export type IteratorsConfiguration<T> = {[Key in keyof T]: IteratorsConfiguration<T[Key]> | Iterator<T[Key]>};